storage,types: refactor time trigger to use interface
