+++ internal/types/time_trigger.go
@@ -1,0 +1,12 @@
+package types
+
+import "time"
+
+type TimeTrigger struct {
+	PolicyID       string
+	CronExpression string
+	StartTime      time.Time
+	EndTime        *time.Time
+	Frequency      string
+	LastExecution  *time.Time
+}
